+++ app/src/main/java/com/talkable/core/view/ViewExt.kt
@@ -1,16 +1,11 @@
 package com.talkable.core.view
 
-import androidx.recyclerview.widget.DiffUtil
+import android.view.View
 
-class ItemDiffCallback<T : Any>(
-    val onItemsTheSame: (T, T) -> Boolean,
-    val onContentsTheSame: (T, T) -> Boolean
-) : DiffUtil.ItemCallback<T>() {
-    override fun areItemsTheSame(
-        oldItem: T, newItem: T
-    ): Boolean = onItemsTheSame(oldItem, newItem)
+fun View.visible(isVisible: Boolean) {
+    this.visibility = if (isVisible) View.VISIBLE else View.GONE
+}
 
-    override fun areContentsTheSame(
-        oldItem: T, newItem: T
-    ): Boolean = onContentsTheSame(oldItem, newItem)
+fun View.toggleSelected() {
+    this.isSelected = !this.isSelected
 }
